https://github.com/e-government-ua/i/issues/1103 minor changes in controller and DAO

+++ wf-base/src/main/java/org/igov/model/action/execute/item/ActionExecuteOldDAOImpl.java
@@ -50,7 +50,7 @@ public class ActionExecuteOldDAOImpl extends GenericEntityDao<ActionExecuteOld>	
 
 	@Transactional
 	@Override
-	public List<ActionExecuteOld> getActionExecute(Integer nRowsMax, String sMethodMask, String asID_Status, Long nTryMax, Long nID) {
+	public List<ActionExecuteOld> getActionExecute(Integer nRowsMax, String sMethodMask, String asID_Status, Integer nTryMax, Long nID) {
 		List<ActionExecuteOld> resList = new ArrayList<ActionExecuteOld>();
 		
 		Criteria criteria = getSession().createCriteria(ActionExecuteOld.class);
